% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Calculation of information criterion}
\usage{
criterion(n, loglik, param)
}
\arguments{
\item{n}{Number of observations}

\item{loglik}{Loglikelihood from model}

\item{param}{Number of parameters fit in model}
}
\value{
A list of the four IC
\itemize{
 \item\code{aic.c}: Akaike information criterion corrected for small samples
 \item\code{aic}: Akaike information criterion
 \item\code{bic}: Bayesian information criterion
 \item\code{hqc}: Hannan-Quinn information criterion
 }
}
\description{
\code{criterion} calculates various information criterion for the algorithms in this package
}
