% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{lidar dataset.}
\format{
A data frame with 221 rows and 2 variables:
\describe{
  \item{range}{distance travelled before the light is reflected back to its source.}
  \item{logratio}{logarithm of the ratio of received light from two laser sources.}
  ...
}
}
\source{
Sigrist, M. (Ed.) (1994). Air Monitoring by Spectroscopic Techniques
(Chemical Analysis Series, vol. 197). New York: Wiley.
}
\usage{
lidar
}
\description{
A dataset containing 221 observations from a light detection and ranging (LIDAR) experiment.
}
\examples{
library(VarReg)
data(lidar)
attach(lidar)
plot(range,logratio)
}
\references{
Ruppert, D., Wand, M.P. and Carroll, R.J. (2003)
Semiparametric Regression Cambridge University Press.
http://stat.tamu.edu/~carroll/semiregbook/
}
\keyword{datasets}
