% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linVarReg.R
\name{linVarReg}
\alias{linVarReg}
\title{Linear mean and variance regression}
\usage{
linVarReg(dat, var.ind = c(2), mean.ind = c(2), para.space = c("all",
  "positive", "negative"), eps = 1e-06, maxit = 1000)
}
\arguments{
\item{dat}{Dataframe containing outcome and covariate data. Outcome data must be in the first column. Covariates for mean and variance model in next columns.}

\item{var.ind}{Vector containing the column numbers of the data in 'dat' to be fit as covariates in the variance model. FALSE indicates constant variance option.}

\item{mean.ind}{Vector containing the column numbers of the data in 'dat' to be fit as covariates in the mean model. 0 indicates constant mean option. NULL indicates zero mean option.}

\item{para.space}{Parameter space to search for variance parameter estimates. "positive" means only search positive parameter space, "negative" means search only negative parameter space and "all" means search all.}

\item{eps}{Very small number for the convergence criteria, default 1 times 10 power6}

\item{maxit}{Number of maximum iterations for the EM algorithm, default 1000.}
}
\value{
$linVarReg$ returns an object of class "VarReg" which inherits some components from the class "glm".

A list of the results from the algorithm, including conv, reldiff, information criterion and mean and variance estimates.
}
\description{
\code{linVarReg} performs multivariate mean and multivariate variance regression.
}
\examples{
data(lidar)
lid<-data.frame(lidar$logratio, lidar$range)
##Fit model with range as a covariate in the mean and the variance model
linmodel<-linVarReg(dat=lidar, var.ind=c(2),mean.ind=c(2), para.space="all")
##Give more iterations to converge
linmodel<-linVarReg(dat=lidar, var.ind=c(2),mean.ind=c(2), maxit=10000, para.space="all")
##Fit model with range as variate in variance model only - constant mean model.
conmean<-linVarReg(dat=lidar, var.ind=c(2),mean.ind=0, para.space="all")
}

