% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarReg.R
\name{plotVarReg}
\alias{plotVarReg}
\title{Plots graphics for a mean and variance regression model}
\usage{
plotVarReg(x, knot.lines = FALSE, CI = FALSE, ci.type = c("im", "boot"),
  xlab = "x", ylab = "y")
}
\arguments{
\item{x}{Object of class VarReg.}

\item{knot.lines}{TRUE/FALSE (Default=TRUE) to show the knot lines on the graphics.}

\item{CI}{TRUE/FALSE (Default=FALSE) to show the 95\% CI on the plots.}
}
\value{
returns a series of graphics.
}
\description{
\code{plotVarReg} to produce graphics for models fit in this package.
}
\examples{
data(lidar)
linmodel<-semiVarReg(lidar$logratio, lidar$range, meanmodel="linear", varmodel="linear")
test<-plotVarReg(linmodel)
}

