% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seVarReg.R
\name{seVarReg}
\alias{seVarReg}
\title{SE for mean and variance regression models}
\usage{
seVarReg(x, boot = FALSE, bootreps = 1000, control = list(...), ...)
}
\arguments{
\item{x}{object of class \code{"VarReg"} to determin the SE (eg. result from \code{\link{"semiVarReg"}}).}

\item{boot}{TRUE or FALSE indicating if bootstrapped CI should be calculated. Default is FALSE.}

\item{bootreps}{Number of bootstraps to be performed if \code{boot=TRUE}. Default is 1000.}

\item{control}{List of control parameters for the bootstrapped models. See \code{\link{VarReg.control}}.}
}
\value{
List of results. \code{mean.est} and \code{variance.est} are tables of results from the mean model and variance model (as appropriate). They include the parameter estimates from the model, SEs from information matrix (if \code{boundary=FALSE}) and if specified, the se from bootstrapping with the bootstrapped 95\% CI. Additionally, the expected information matrices are given for the mean and the variance (\code{mean.im} and \code{variance.im}).

Need to work on the inclusion of a vector of x values, spit out a vector of se for those specified x values.
}
\description{
\code{seVarReg} calculates SE for an object of class \code{"VarReg"}. If the result is not on a boundary, the Fishers Information matrix SE are given. The bootstrapped 95\% CI can also be calculated.
}
\examples{
data(lidar)
lid<-data.frame(lidar$logratio, lidar$range)
##Fit model with range as a covariate in the mean and the variance model
semimodel<-semiVarReg(lidar$logratio, lidar$range, meanmodel="semi", varmodel="semi", knots.m=4, knots.v=2)
##Calculate SE
se<-seVarReg(semimodel, boot=TRUE, bootreps=10)
}

