% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiVarReg.R
\name{semiVarReg}
\alias{semiVarReg}
\title{Semi parametric mean and variance regression}
\usage{
semiVarReg(y, x, meanmodel = c("zero", "constant", "linear", "semi"),
  varmodel = c("constant", "linear", "semi"), knots.m = 2, knots.v = 2,
  degree = 2, mono.var = c("none", "inc", "dec"), control = list(...),
  ...)
}
\arguments{
\item{y}{Vector containing outcome data. Must be no missing data.}

\item{x}{Vector containing the covariate data. Must be no missing data.}

\item{meanmodel}{Specify the mean model to be fit: "zero" =zero mean model, "constant" = constant mean, "linear" = linear term with x covariate, "semi" = semiparametric spline (specify with \code{knots.m}).}

\item{varmodel}{Specify the variance model to be fit: "constant" = constant variance model, "linear" = linear term with x covariate, "semi" = semiparametric spline (specify with \code{knots.v})}

\item{knots.m}{A integer indicating the number of internal knots to be fit in the mean model. Default is '2'. (Note that the knots are placed equidistantly over x.)}

\item{knots.v}{A integer indicating the number of internal knots to be fit in the variance model. Default is '2'. (Note that the knots are placed equidistantly over x.)}

\item{degree}{The degree of the splines fit in the mean and variance. Default is '2'.}

\item{mono.var}{Indicates whether the variance model is monotonic (only applicable for 'linear' or 'semi' variance models). Default is 'none' (no monotonic constraints). Options are 'inc' for increasing or 'dec' for decreasing.}

\item{control}{list of control parameters. See \code{\link{VarReg.control}}.}
}
\value{
\code{semiVarReg} returns an object of class "\code{VarReg}" which inherits some components from the class "\code{glm}".

Includes converge, reldiff, information criterion and mean and variance estimates.
}
\description{
\code{semiVarReg} performs a semiparametric mean and variance regression model. Currently, this is only designed for a single covariate that is fit in the mean and variance models.
}
\examples{
data(lidar)
linmodel<-semiVarReg(lidar$logratio, lidar$range, meanmodel="linear", varmodel="linear")
semimodel<-semiVarReg(lidar$logratio, lidar$range, meanmodel="semi", varmodel="semi", knots.m=4, knots.v=2)
}

