% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seVarReg.R
\name{seVarReg}
\alias{seVarReg}
\title{SE for mean and variance regression}
\usage{
seVarReg(x, boot = FALSE, bootreps = 1000, eps = 1e-06, maxit = 1000)
}
\arguments{
\item{x}{object of class VarReg to determin the SE (eg. model from semiVarReg)}

\item{boot}{TRUE or FALSE indicating if bootstrapped CI should be calculated. Default is FALSE.}

\item{bootreps}{Number of bootstraps to be performed if $boot=TRUE$. Default is 1000.}
}
\value{
Table of results.
}
\description{
\code{seVarReg} calculates SE for an object from a linVarReg. If the result is not on a boundary, the Fishers Information matrix SE are given. The bootstrapped 95% CI can also be calculated.
}
\examples{
data(lidar)
lid<-data.frame(lidar$logratio, lidar$range)
##Fit model with range as a covariate in the mean and the variance model
semimodel<-semiVarReg(lidar$logratio, lidar$range, meanmodel="semi", varmodel="semi", knots.m=4, knots.v=2)
##Calculate SE
se<-seVarReg(semimodel, boot=TRUE, bootreps=10)
}

