% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarReg.R
\docType{package}
\name{VarReg}
\alias{VarReg}
\alias{VarReg-package}
\title{VarReg: A package for multivariate mean and variance regression models}
\description{
The  package provides  important functions:
}
\section{functions}{

The  functions ...
}

